Update ChocolateyInstall.ps1
I've been getting "failures" even though this package has installed correctly. The return code is 5100, which most of the web agrees means that the system needs a restart, the difference between this and 3010 appears to be that 5100 means it was in a "requires reboot" state BEFORE the current package was installed, where 3010 means that "I triggered the reboot requirement".