chocolatey-windowsupdate.extension: treat WU_E_NOT_APPLICABLE as a success
When an older update has been superseded by a newer one, this exit code
may appear on systems which only have the newer update installed, even
though Test-WindowsUpdate will return False. This should be treated as a
success.

This exit code may also appear if a prerequisite update or Service Pack is
missing. It is the responsibility of the package maintainer to ensure all
such prerequisites are encoded in the package, either as dependencies or
as service pack rules.