chocolatey-windowsupdate.extension: work around a bug in Set-PowerShellExitCode
The current implementation of Set-PowerShellExitCode incorrectly compares
an int parameter with null - this comparison will be true if the parameter
is 0, because null will be coerced to int (0). This error makes that
function ignore ExitCode 0.

Provide a shim over Set-PowerShellExitCode, which detects this case and
corrects it.

Although this code uses assumptions about the inner workings of
Set-PowerShellExitCode, it should be reasonably future-proof (unless
Chocolatey starts using the ChocolateyExitCode environment variable for a
different purpose).