dotnet4.6.2, netfx-4.6.2-devpack: fix installation on Windows Server Core
The .NET 4.6.2 installer crashes on Server Core in passive mode, because
it attempts to use a small piece of GUI functionality that is unavailable
on Core (same as .NET 4.6 and 4.6.1 [1]).

Switch to quiet mode to fix this.

The DevPack package depends on the Runtime package, so the dependency
needs to be updated. Although the DevPack installer works fine in passive
mode on Core, switch it to quiet for symmetry, too.

Fixes #3.

[1] http://serverfault.com/questions/733596/unable-to-install-net-framework-4-6-on-windows-server-2012-r2-core