visualstudio2017buildtools: update dependencies
KB3033929 is not needed (unlike Enterprise).
KB2999226 is installed on Win 8.1 by the core product and on Win 7 by the
VCTools workload.
The installer requires .NET 4.6 to start, but installs 4.6.1 while
installing the core product.

vcredist2017 is only required by the VCTools workload - to be moved
ultimately to the workload package.