chocolatey-windowsupdate.extension: handle WU_S_ALREADY_INSTALLED
In some circumstances an update may not be exposed as a
Win32_QuickFixEngineering instance, but may still be considered installed
by the system. This might probably happen if the update has been
superseded by another already installed update and for some reason the
system does not return WU_E_NOT_APPLICABLE.

Fixes #5.