chocolatey-visualstudio.extension: allow skipping VS Installer update
This may be useful when instaling a fresh instance of a VS product using
a bootstrapper which installs a specific version of the product
(e.g. the 15.0 bootstrappers available from MSDN).

Note that because the 15.0 bootstrapper does not support
--noUpdateInstaller, the package must be instructed not to pass this
parameter to the bootstrapper, by additionally specifying
--no-noUpdateInstaller.

GitHub-Issue: GH-7 GH-8 GH-10 GH-26