chocolatey-visualstudio.extension: work around broken --update feature of the bootstrapper
Due to the way the VS Installer update feature of the VS setup
bootstrapper (--update) is implemented, --update must be the last
argument, otherwise the bootstrapper will assume it is in "roundtrip
update" mode and will start the VS Installer once it is updated.

The problem is that recent versions of the parent ("box") application
(vs_<product>.exe) started appending an additional argument (--env) to
the VS setup bootstrapper invocation, effectively breaking the --update
feature. To add insult to injury, the --env argument is not used during
the VS installer update.

Fortunately, the box application is just a flavor of a self-extracting
zip archive and can be unzipped using standard zip tools. So, the
workaround is to unzip the box application and invoke the extracted VS
setup bootstrapper, taking care to pass arguments in correct order.

GitHub-Issue: GH-7 GH-8 GH-26