Exit script if update doesn't apply to OS
Prevent Install-WindowsUpdate being called if the update doesn't apply to the OS version. 
Note: Installing the Visual Studio 2017 package on Windows Server 2016 / Windows 10 will cause this package to be installed as well (even though we don't need it), because it's a dependency. When using Boxstarter to do this, Install-WindowsUpdate will clash with Install-WindowsUpdate from Boxstarter, which then goes ahead and installs ALL critical updates. This scoping issue is more a problem for Boxstarter to fix, but in the mean time, I'd like to work around this issue by preventing Install-WindowsUpdate being called at all if the update doesn't apply to the OS.