chocolatey-visualstudio.extension: fix installation when current location is a network share
The PowerShell concept of current location is in principle much broader
than the filesystem (due to the provider concept), so the current PowerShell
location does not need to be a filesystem location.
Even in filesystem case, there are types of locations unacceptable as
working directory for a process, such as locations on network shares.

Avoid all those issues by specifying a well defined, known to exist
location.

Fixes GH-59.