Added IncludeOptionalComponentsByDefault switch
I install a visual studio build tools package as a dependency, which cant specify install parameters. So ive created a wrapper package that adds the optional flag. Right now as a workaround Ive copied most of this method and invoke Start-VSModifyOperation manually, but with this change that code would be cleaner.