chocolatey-visualstudio.extension: update VS Installer to the highest version required by installed VS products
The VS Installer checks all channels of already installed VS products to
determine if it needs to be updated, not only the channel of the product
being currently installed/modified. The extension should do that, too,
to maintain control over the VS Installer processes (when the VS
Installer self-updates, it does not respect --wait).

The extension will not check the other channels if --noWeb is passed in
package parameters (to avoid going online). Channel uris stored in the
installed VS product metadata are used (so the same that the VS
Installer would use). If a VS product was installed with updates turned
off (by passing a blank channelUri to the installer), it is skipped.

As before, passing --noUpdateInstaller turns off all VS Installer
update logic in the extension.

Related issue: GH-97